Isolate provider descriptor from AC
